src/cache.py
"""
Redis-based hot vector cache implementation.

This module provides the in-memory cache layer for MemVec, handling:
- Connection to Redis server
- Vector storage and retrieval with TTL
- LRU/LFU eviction policies
- Multi-get operations for batch vector lookups
"""

import redis
import numpy as np
from typing import List, Optional, Dict
import os
from dotenv import load_dotenv

load_dotenv()


class VectorCache:
    """Redis-based vector cache for hot storage."""
    
    def __init__(self):
        """Initialize Redis connection."""
        # TODO: Implement Redis connection setup
        pass
    
    def get_vector(self, vector_id: str) -> Optional[np.ndarray]:
        """Retrieve a single vector by ID."""
        # TODO: Implement single vector retrieval
        pass
    
    def get_vectors(self, vector_ids: List[str]) -> Dict[str, np.ndarray]:
        """Retrieve multiple vectors by IDs (multi-get)."""
        # TODO: Implement batch vector retrieval
        pass
    
    def put_vector(self, vector_id: str, vector: np.ndarray, ttl: int = 3600):
        """Store a vector in cache with TTL."""
        # TODO: Implement vector storage with expiration
        pass
    
    def put_vectors(self, vectors: Dict[str, np.ndarray], ttl: int = 3600):
        """Store multiple vectors in cache with TTL."""
        # TODO: Implement batch vector storage
        pass
    
    def evict(self, vector_id: str) -> bool:
        """Manually evict a vector from cache."""
        # TODO: Implement manual eviction
        pass
    
    def clear(self):
        """Clear all cached vectors."""
        # TODO: Implement cache clearing
        pass
