src/index.py
"""
FAISS-based ANN index for candidate selection.

This module provides the approximate nearest neighbor search functionality:
- FAISS index initialization and management
- Vector indexing for fast similarity search
- Candidate ID retrieval for queries
- Index persistence and loading
"""

import faiss
import numpy as np
from typing import List, Tuple
import os


class ANNIndex:
    """FAISS-based ANN index for fast candidate selection."""
    
    def __init__(self, dimension: int, index_type: str = "IVF"):
        """Initialize FAISS index with specified dimension and type."""
        # TODO: Implement FAISS index initialization
        self.dimension = dimension
        self.index_type = index_type
        self.index = None
        pass
    
    def build_index(self, vectors: np.ndarray, vector_ids: List[str]):
        """Build the ANN index from vectors and their IDs."""
        # TODO: Implement index building from vector data
        pass
    
    def add_vectors(self, vectors: np.ndarray, vector_ids: List[str]):
        """Add new vectors to existing index."""
        # TODO: Implement incremental vector addition
        pass
    
    def search(self, query_vector: np.ndarray, k: int = 200) -> Tuple[List[str], List[float]]:
        """Search for k nearest candidate IDs and their distances."""
        # TODO: Implement ANN search returning candidate IDs and scores
        pass
    
    def save_index(self, filepath: str):
        """Save the index to disk."""
        # TODO: Implement index persistence
        pass
    
    def load_index(self, filepath: str):
        """Load index from disk."""
        # TODO: Implement index loading
        pass
    
    def get_stats(self) -> dict:
        """Get index statistics (size, vectors count, etc.)."""
        # TODO: Implement index statistics
        pass
