requirements.txt
faiss-cpu
redis
boto3
python-dotenv
numpy

# Development dependencies
pytest
black
flake8
mypy
